move other two utils
