+++ datafusion/physical-expr-core/src/aggregate/utils.rs
@@ -17,6 +17,13 @@
 
 use std::{any::Any, sync::Arc};
 
+use arrow::{
+    compute::SortOptions,
+    datatypes::{DataType, Field},
+};
+
+use crate::sort_expr::PhysicalSortExpr;
+
 use super::AggregateExpr;
 
 /// Downcast a `Box<dyn AggregateExpr>` or `Arc<dyn AggregateExpr>`
@@ -34,4 +41,29 @@ pub fn down_cast_any_ref(any: &dyn Any) -> &dyn Any {
     } else {
         any
     }
+}
+
+/// Construct corresponding fields for lexicographical ordering requirement expression
+pub fn ordering_fields(
+    ordering_req: &[PhysicalSortExpr],
+    // Data type of each expression in the ordering requirement
+    data_types: &[DataType],
+) -> Vec<Field> {
+    ordering_req
+        .iter()
+        .zip(data_types.iter())
+        .map(|(sort_expr, dtype)| {
+            Field::new(
+                sort_expr.expr.to_string().as_str(),
+                dtype.clone(),
+                // Multi partitions may be empty hence field should be nullable.
+                true,
+            )
+        })
+        .collect()
+}
+
+/// Selects the sort option attribute from all the given `PhysicalSortExpr`s.
+pub fn get_sort_options(ordering_req: &[PhysicalSortExpr]) -> Vec<SortOptions> {
+    ordering_req.iter().map(|item| item.options).collect()
 }
